folder restructuring - closes #48
