+++ bin/createsrcdirandfiles.js
@@ -5,10 +5,26 @@
 const fs = require("fs");
 
 const openAppendFile = require("./openandappendfile");
-const dataInFiles = require("../tasks/data");
-const sequelizeSetupData = require("../tasks/database/sequelizedata");
-const userModelData = require("../tasks/database/model");
-const { controllers, userController } = require("../tasks/controllers/user");
+const {
+    controllers,
+    userController,
+    userModelData,
+    noSequelizeSetupData,
+    sequelizeSetupData,
+    middleware,
+    userMiddleware,
+    noSequelizeUserModelData,
+    routes,
+    userRouter,
+    appJsTest,
+    createDb,
+    dropDb,
+    noOrmcreateDb,
+    noOrmDropDb,
+    userQueries,
+    appJs,
+    noOrmUserController
+} = require("../tasks");
 
 const createSrcDirAndFiles = (details) => {
     const { appBaseDirectory, tests, database, orm } = details;
@@ -18,6 +34,7 @@ const createSrcDirAndFiles = (details) => {
     }
 
     const folders = foldersToAdd.map(folder => {
+        // will need to recheck here
         if (folder || folder !== undefined) {
             return `${appBaseDirectory}/${folder}`
         }
@@ -43,16 +60,16 @@ const createSrcDirAndFiles = (details) => {
                 const dropDbFileName = fs.createWriteStream(dropDbFile);
                 let createDbData, dropDbData;
                 if (orm && orm.toLowerCase() === 'sequelize') {
-                    createDbData = dataInFiles.createDb;
-                    dropDbData = dataInFiles.dropDb;
+                    createDbData = createDb;
+                    dropDbData = dropDb;
                 } else if (orm && orm.toLowerCase() === 'no orm') {
                     // When user does't need any orm
-                    createDbData = dataInFiles.noOrmcreateDb;
-                    dropDbData = dataInFiles.noOrmDropDb;
+                    createDbData = noOrmcreateDb;
+                    dropDbData = noOrmDropDb;
                     // Create a script to query database -> only when no orm
                     const queriesFile = `${appBaseDirectory}/src/scripts/queries.js`;
                     const queriesFileName = fs.createWriteStream(queriesFile);
-                    openAppendFile(queriesFileName.path, dataInFiles.userQueries);
+                    openAppendFile(queriesFileName.path, userQueries);
                 }
                 // append drop database scripts and create db scripts
                 openAppendFile(createDbFileName.path, createDbData);
@@ -84,9 +101,9 @@ const createSrcDirAndFiles = (details) => {
                          * In case sqlite set in the future so that it does'nt make use of orm
                          * then, we can have another if statement for the same here
                          */
-                        openAppendFile(pathName, dataInFiles.noSequelizeSetupData);
+                        openAppendFile(pathName, noSequelizeSetupData);
                         // Add data to user.js file when no orm is chosen in models 
-                        openAppendFile(userModelsFileName.path, dataInFiles.noSequelizeUserModelData);
+                        openAppendFile(userModelsFileName.path, noSequelizeUserModelData);
                     }
                 }
             }
@@ -94,13 +111,13 @@ const createSrcDirAndFiles = (details) => {
             // Write to test file.
             if (fileName.path === `${appBaseDirectory}/test/index.js`) {
                 const pathName = fileName.path;
-                openAppendFile(pathName, dataInFiles.appJsTest);
+                openAppendFile(pathName, appJsTest);
             };
 
             // Add data to the main entry point
             if (fileName.path === `${appBaseDirectory}/src/index.js`) {
                 const pathName = fileName.path;
-                openAppendFile(pathName, dataInFiles.appJs);
+                openAppendFile(pathName, appJs);
             };
 
             // Paths that needs files such as user.js, config may not need it.
@@ -118,14 +135,14 @@ const createSrcDirAndFiles = (details) => {
                 const directoryFiles = `${folder}/user.js`;
                 const directoryFileName = fs.createWriteStream(directoryFiles);
                 if (directoryFileName.path === `${appBaseDirectory}/src/middlewares/user.js`) {
-                    openAppendFile(directoryFileName.path, dataInFiles.userMiddleware);
+                    openAppendFile(directoryFileName.path, userMiddleware);
                 }
                 if (directoryFileName.path === `${appBaseDirectory}/src/routes/user.js`) {
-                    openAppendFile(directoryFileName.path, dataInFiles.userRouter);
+                    openAppendFile(directoryFileName.path, userRouter);
                 }
                 if (directoryFileName.path === `${appBaseDirectory}/src/controllers/user.js`) {
                     if (orm === 'No ORM') {
-                        openAppendFile(directoryFileName.path, dataInFiles.noOrmUserController);
+                        openAppendFile(directoryFileName.path, noOrmUserController);
                     } else {
                         openAppendFile(directoryFileName.path, userController);
                     }
@@ -134,10 +151,10 @@ const createSrcDirAndFiles = (details) => {
 
             // Write to different files including the index.js and other files
             if (fileName.path === `${appBaseDirectory}/src/middlewares/index.js`) {
-                openAppendFile(fileName.path, dataInFiles.middleware);
+                openAppendFile(fileName.path, middleware);
             }
             if (fileName.path === `${appBaseDirectory}/src/routes/index.js`) {
-                openAppendFile(fileName.path, dataInFiles.routes);
+                openAppendFile(fileName.path, routes);
             }
             if (fileName.path === `${appBaseDirectory}/src/controllers/index.js`) {
                 openAppendFile(fileName.path, controllers);
