+++ tasks/config/babel.js
@@ -1,0 +1,15 @@
+
+const babel = `{
+    "presets": [
+        [
+            "@babel/preset-env",
+            {
+                "targets": {
+                    "node": "current"
+                }
+            }
+        ]
+    ]
+}`;
+
+exports = module.exports = babel;
