+++ tasks/config/envexample.js
@@ -1,0 +1,7 @@
+const envExample = `
+TEST_DATABASE_URL="postgres://localhost:5432/dbnameTest"
+DATABASE_URL="postgres://localhost:5432/dbname"
+NODE_ENV="development"
+`;
+
+exports = module.exports = envExample;
