+++ tasks/controllers/index.js
@@ -1,0 +1,3 @@
+const controllers = `export { default as UserController } from './user';`
+
+exports = module.exports = controllers;
