+++ tasks/database/noormdropdb.js
@@ -1,0 +1,9 @@
+const noOrmDropDb = `import db from '../models/setup';
+db.query('drop database users', (err, result) => {
+    if (err) {
+        console.log("The Error", error);
+    }
+    console.log('Successfully dropped db')
+}); `;
+
+exports = module.exports = noOrmDropDb;
