+++ tasks/middlewares/index.js
@@ -1,0 +1,3 @@
+const middleware = `export {default as UserMiddleware} from './user';`
+
+exports = module.exports = middleware;
