+++ tasks/middlewares/user.js
@@ -1,0 +1,6 @@
+const userMiddleware = `export default function userMiddleware(req, res, next) {
+    next();
+}
+`;
+
+exports = module.exports = userMiddleware;
