+++ tasks/models/sequelizeusermodel.js
@@ -55,5 +55,4 @@ const userModelData = (database) => {
     )
 }
 
-
 exports = module.exports = userModelData;
